Add files via upload
This is a PDF file containing a presentation that I did for a final project in an IoT class in Spring 2024. This final involved using a Raspberry Pi as an Ansible Controller.
Additionally, I used a simple Python Script in this project to take sensor data (external and internal) and then send that data to an online dashboard. 
This project used the Ansible Version Control repository for the Ansible portion.